+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/PointColor/PointColorItem.tsx
@@ -1,9 +1,9 @@
 import { useCallback } from "react";
 
-import { Cond } from "../types";
+import { ColorField, ConditionField, ItemControls } from "../../../common";
+import { Cond } from "../../types";
 
-import { ColorField, ConditionField, ItemControls } from "./common";
-import { Item } from "./commonComponents";
+import { Item } from "./../../../commonComponents";
 
 const PointColorItem: React.FC<{
   index: number;
