+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/PointColor/index.tsx
@@ -2,9 +2,9 @@ import AddButton from "@web/extensions/sidebar/core/components/content/common/Da
 import { generateID, moveItemDown, moveItemUp, removeItem } from "@web/extensions/sidebar/utils";
 import { useCallback, useState } from "react";
 
-import { BaseFieldProps, Cond } from "../types";
+import { ButtonWrapper, Wrapper } from "../../../commonComponents";
+import { BaseFieldProps, Cond } from "../../types";
 
-import { ButtonWrapper, Wrapper } from "./commonComponents";
 import PointColorItem from "./PointColorItem";
 
 const PointColor: React.FC<BaseFieldProps<"pointColor">> = ({ value, editMode, onUpdate }) => {
