+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/PointLabel.tsx
@@ -1,10 +1,10 @@
 import { styled } from "@web/theme";
 import { ChangeEvent, useCallback, useState } from "react";
 
+import { ColorField, SelectField, SwitchField, TextField } from "../../common";
 import { BaseFieldProps, Fields } from "../types";
 
-import { ColorField, SelectField, SwitchField, TextField } from "./common";
-import { Wrapper } from "./commonComponents";
+import { Wrapper } from "./../../commonComponents";
 
 // TODO: list all options for select field dropdown
 const options = [
