+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/PointStroke/index.tsx
@@ -2,9 +2,9 @@ import AddButton from "@web/extensions/sidebar/core/components/content/common/Da
 import { generateID, moveItemDown, moveItemUp, removeItem } from "@web/extensions/sidebar/utils";
 import { useCallback, useState } from "react";
 
-import { BaseFieldProps, Cond } from "../types";
+import { ButtonWrapper, Wrapper } from "../../../commonComponents";
+import { BaseFieldProps, Cond } from "../../types";
 
-import { ButtonWrapper, Wrapper } from "./commonComponents";
 import PointStrokeItem from "./PointStrokeItem";
 
 const PointStroke: React.FC<BaseFieldProps<"pointStroke">> = ({ value, editMode, onUpdate }) => {
