+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/common/ColorField.tsx
